// src/app/public/shared/components/filters/core/enums/configuration-ids/car-configuration-ids.enum.ts
export enum CAR_CONFIGURATION_ID {
  PRICE = 'cars-price',
  BRAND_N_MODEL = 'cars-brand-n-model',
  YEAR = 'cars-year',
  KM = 'cars-km',
  BODY = 'cars-body',
  ENGINE = 'cars-engine',
  GEARBOX = 'cars-gearbox',
  WARRANTY = 'cars-warranty',
  SEATS = 'cars-seats',
  DOORS = 'cars-doors',
  HORSE_POWER = 'cars-horse-power',
  PROFESSIONAL = 'cars-professional',
}